fix: `sed -i` discrepancy between macOS and Linux
Problem:
Not supplying an extension to the `sed -i` command causes macOS to com-
plain, while it's okay on Linux machines. Doing the inverse fixes mac-
OS but then Linux machines will complain.

Solution:
Checking the user's OS and supplying the empty extension if they are on
macOS and not supplying and extension to the command on Linux seems to
solve the issue.

Closes: #2